Updated TR's stat tracker to include long int Update CHANGELOG.md Removed unnecessary new lines

+++ traffic_router/core/src/main/java/com/comcast/cdn/traffic_control/traffic_router/core/router/StatTracker.java
@@ -318,14 +318,14 @@ public class StatTracker {
 	public Map<String, Tallies> getHttpMap() {
 		return httpMap;
 	}
-	public int getTotalDnsCount() {
+	public long getTotalDnsCount() {
 		return totalDnsCount;
 	}
 	public long getAverageDnsTime() {
 		if(totalDnsCount==0) { return 0; }
 		return totalDnsTime/totalDnsCount;
 	}
-	public int getTotalHttpCount() {
+	public long getTotalHttpCount() {
 		return totalHttpCount;
 	}
 	public long getAverageHttpTime() {
@@ -339,9 +339,9 @@ public class StatTracker {
 		this.totalDsMissCount = totalDsMissCount;
 	}
 
-	private int totalDnsCount;
+	private long totalDnsCount;
 	private long totalDnsTime;
-	private int totalHttpCount;
+	private long totalHttpCount;
 	private long totalHttpTime;
 	private int totalDsMissCount = 0;
 	public Map<String,Long> getUpdateTracker() {
